Moved to Sets for solver
