+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/CommonDenominatorSolver.cs
@@ -59,7 +59,7 @@ namespace AngouriMath.Functions.Algebra.AnalyticalSolving
         /// Finds the best common denominator, multiplies the whole expression by that, and
         /// tries solving if the found denominator is not 1
         /// </summary>
-        internal static IEnumerable<Entity>? Solve(Entity expr, Variable x) =>
+        internal static SetNode? Solve(Entity expr, Variable x) =>
             FindCD(expr, x) is { } res ? AnalyticalSolver.Solve(res, x) : null;
 
         private static Entity? FindCD(Entity expr, Variable x)
