+++ AngouriMath/Functions/Continuous/Algebra/InvertNode.Definition.cs
@@ -1,0 +1,55 @@
+﻿
+/* Copyright (c) 2019-2020 Angourisoft
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
+ * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software
+ * is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+using static AngouriMath.Entity.Number;
+using System.Collections.Generic;
+using System.Linq;
+using AngouriMath.Core;
+
+namespace AngouriMath
+{
+    public abstract partial record Entity : ILatexiseable
+    {
+        /// <summary><para>This <see cref="Entity"/> MUST contain exactly ONE occurance of <paramref name="x"/>,
+        /// otherwise this function won't work correctly.</para>
+        /// 
+        /// This function inverts an expression and returns a <see cref="Set"/>. Here, a represents <paramref name="value"/>.
+        /// <list type="table">
+        /// <item>x^2 = a ⇒ x = { sqrt(a), -sqrt(a) }</item>
+        /// <item>sin(x) = a ⇒ x = { arcsin(a) + 2 pi n, pi - arcsin(a) + 2 pi n }</item>
+        /// </list>
+        /// </summary>
+        /// <returns>A set of possible roots of the expression.</returns>
+        internal IEnumerable<Entity> Invert(Entity value, Entity x) =>
+            value.InnerSimplify() is var simplified && this == x
+            ? new[] { simplified }
+            : InvertNode(simplified, x).Where(el => el.IsFinite);
+        /// <summary>Use <see cref="Invert(Entity, Entity)"/> instead which auto-simplifies <paramref name="value"/></summary>
+        private protected abstract IEnumerable<Entity> InvertNode(Entity value, Entity x);
+        /// <summary>
+        /// Returns true if <paramref name="a"/> is inside a rect with corners <paramref name="from"/>
+        /// and <paramref name="to"/>, OR <paramref name="a"/> is an unevaluable expression
+        /// </summary>
+        private protected static bool EntityInBounds(Entity a, Complex from, Complex to)
+        {
+            if (!(a.Evaled is Complex r))
+                return true;
+            return r.RealPart >= from.RealPart &&
+                   r.ImaginaryPart >= from.ImaginaryPart &&
+                   r.RealPart <= to.RealPart &&
+                   r.ImaginaryPart <= to.ImaginaryPart;
+        }
+    }
+}
